feat: DTO e MAPPER configurados para MissoesController e MissoesModel, ResponseEntity adicionado funcionando e devolvendo as mensagens para o usuario

+++ src/main/java/dev/java10x/cadastrodeninja/Missoes/Controller/MissoesController.java
@@ -1,9 +1,12 @@
 package dev.java10x.cadastrodeninja.Missoes.Controller;
 
-import dev.java10x.cadastrodeninja.Missoes.Model.MissoesModel;
+import dev.java10x.cadastrodeninja.Missoes.DTOMAPPERMISSOES.MissoesDTO;
+
 import dev.java10x.cadastrodeninja.Missoes.Service.MissoesService;
-import dev.java10x.cadastrodeninja.Ninjas.Model.NinjaModel;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -12,34 +15,55 @@ import java.util.List;
 @RequestMapping("/missoes")
 public class MissoesController {
 
+    @Autowired
     private MissoesService missoesService;
 
-    public MissoesController(MissoesService missoesService) {
-        this.missoesService = missoesService;
-    }
-
     @GetMapping("/mostrar")
-    public List<MissoesModel> mostrar(){
-        return missoesService.mostrar();
+    public ResponseEntity<List<MissoesDTO>> mostrar() {
+        List<MissoesDTO> missoes = missoesService.mostrar();
+        return ResponseEntity.ok(missoes);
     }
 
     @GetMapping("/mostrar/{id}")
-    public MissoesModel mostrar(@PathVariable Long id){
-        return missoesService.mostrarPorID(id);
+    public ResponseEntity<?> mostrarPorID(@PathVariable Long id) {
+        MissoesDTO missaoPorId = missoesService.mostrarPorID(id);
+        if (missaoPorId != null) {
+            return ResponseEntity.ok(missaoPorId);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("A missao com o ID " + id + " Nao foi encontrada");
+        }
     }
 
     @PostMapping("/criarMissao")
-    public MissoesModel adicionarMissao(@RequestBody MissoesModel missao) {
-        return missoesService.adicionarMissao(missao);
+    public ResponseEntity<String> adicionarMissao(@RequestBody MissoesDTO missaoDTO) {
+        MissoesDTO NovasMissoes = missoesService.adicionarMissao(missaoDTO);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Missao criada com sucesso: " + NovasMissoes.getNome() + "(ID): " + NovasMissoes.getId());
     }
 
     @PutMapping("/alterarMissao/{id}")
-    public MissoesModel alterarMissaoID(@PathVariable Long id, @RequestBody MissoesModel missao) {
-        return missoesService.atualizarMissao(id, missao);
+    public ResponseEntity<?> alterarMissaoID(@PathVariable Long id, @RequestBody MissoesDTO missaoDTO) {
+
+        MissoesDTO missaoAlterada = missoesService.atualizarMissao(id, missaoDTO);
+
+        if (missaoAlterada != null) {
+            return ResponseEntity.ok("Missao alterada com sucesso" + missaoAlterada);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Missao com o id: " + id + " Nao foi encontrada");
+        }
     }
 
+
     @DeleteMapping("/deletarMissao/{id}")
-    public void DeletarMissaoID(@PathVariable Long id) {
-        missoesService.deleteMissionById(id);
+    public ResponseEntity<String> DeletarMissaoID(@PathVariable Long id) {
+        if (missoesService.mostrarPorID(id) != null) {
+            missoesService.deleteMissionById(id);
+            return ResponseEntity.ok("Missao deletada com sucesso!" + id);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("A missao com o id " + id + " nao foi encontrada");
+        }
     }
 }
